@@ -6,24 +6,25 @@ import 'dart:typed_data';
 
 import 'package:file/memory.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'dart:async';
 
 class FakeCacheManager extends Mock implements CacheManager {
   void throwsNotFound(String url) {
-    when(getFileStream(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    )).thenThrow(HttpExceptionWithStatus(404, 'Invalid statusCode: 404',
-        uri: Uri.parse(url)));
+    when(() => getFileStream(
+              url,
+              key: any(named: 'key'),
+              headers: any(named: 'headers'),
+              withProgress: any(named: 'withProgress'),
+            ))
+        .thenThrow(HttpExceptionWithStatus(404, 'Invalid statusCode: 404',
+            uri: Uri.parse(url)));
   }
 
   ExpectedData returns(
     String url,
     List<int> imageData, {
-    Duration delayBetweenChunks,
+    Duration? delayBetweenChunks,
   }) {
     const chunkSize = 8;
     final chunks = <Uint8List>[
@@ -31,12 +32,12 @@ class FakeCacheManager extends Mock implements CacheManager {
         Uint8List.fromList(imageData.skip(offset).take(chunkSize).toList()),
     ];
 
-    when(getFileStream(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    )).thenAnswer((realInvocation) => _createResultStream(
+    when(() => getFileStream(
+          url,
+          key: any(named: 'key'),
+          headers: any(named: 'headers'),
+          withProgress: any(named: 'withProgress'),
+        )).thenAnswer((_) => _createResultStream(
           url,
           chunks,
           imageData,
@@ -54,7 +55,7 @@ class FakeCacheManager extends Mock implements CacheManager {
     String url,
     List<Uint8List> chunks,
     List<int> imageData,
-    Duration delayBetweenChunks,
+    Duration? delayBetweenChunks,
   ) async* {
     var totalSize = imageData.length;
     var downloaded = 0;
@@ -74,27 +75,29 @@ class FakeCacheManager extends Mock implements CacheManager {
 
 class FakeImageCacheManager extends Mock implements ImageCacheManager {
   ExpectedData returns(
-      String url,
-      List<int> imageData, {
-        Duration delayBetweenChunks,
-      }) {
+    String url,
+    List<int> imageData, {
+    Duration? delayBetweenChunks,
+  }) {
     const chunkSize = 8;
     final chunks = <Uint8List>[
       for (int offset = 0; offset < imageData.length; offset += chunkSize)
         Uint8List.fromList(imageData.skip(offset).take(chunkSize).toList()),
     ];
 
-    when(getImageFile(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    )).thenAnswer((realInvocation) => _createResultStream(
-      url,
-      chunks,
-      imageData,
-      delayBetweenChunks,
-    ));
+    when(() => getImageFile(
+          url,
+          key: any(named: 'key'),
+          headers: any(named: 'headers'),
+          withProgress: any(named: 'withProgress'),
+          maxHeight: any(named: 'maxHeight'),
+          maxWidth: any(named: 'maxWidth'),
+        )).thenAnswer((_) => _createResultStream(
+          url,
+          chunks,
+          imageData,
+          delayBetweenChunks,
+        ));
 
     return ExpectedData(
       chunks: chunks.length,
@@ -104,11 +107,11 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
   }
 
   Stream<FileResponse> _createResultStream(
-      String url,
-      List<Uint8List> chunks,
-      List<int> imageData,
-      Duration delayBetweenChunks,
-      ) async* {
+    String url,
+    List<Uint8List> chunks,
+    List<int> imageData,
+    Duration? delayBetweenChunks,
+  ) async* {
     var totalSize = imageData.length;
     var downloaded = 0;
     for (var chunk in chunks) {
@@ -130,5 +133,9 @@ class ExpectedData {
   final int totalSize;
   final int chunkSize;
 
-  const ExpectedData({this.chunks, this.totalSize, this.chunkSize});
+  const ExpectedData({
+    required this.chunks,
+    required this.totalSize,
+    required this.chunkSize,
+  });
 }
