@@ -11,11 +11,11 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mocktail/mocktail.dart';
 
 import 'fake_cache_manager.dart';
 import 'image_data.dart';
 import 'rendering_tester.dart';
-import 'package:mockito/mockito.dart';
 
 void main() {
   TestRenderingFlutterBinding();
@@ -23,8 +23,8 @@ void main() {
   setUp(() {});
 
   tearDown(() {
-    PaintingBinding.instance.imageCache.clear();
-    PaintingBinding.instance.imageCache.clearLiveImages();
+    PaintingBinding.instance?.imageCache?.clear();
+    PaintingBinding.instance?.imageCache?.clearLiveImages();
   });
 
   test('Supplying an ImageCacheManager should call getImageFile', () async {
@@ -45,19 +45,21 @@ void main() {
     ));
     await imageAvailable.future;
 
-    verify(cacheManager.getImageFile(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    )).called(1);
+    verify(() => cacheManager.getImageFile(
+          url,
+          key: any(named: 'key'),
+          headers: any(named: 'headers'),
+          withProgress: any(named: 'withProgress'),
+          maxHeight: any(named: 'maxHeight'),
+          maxWidth: any(named: 'maxWidth'),
+        )).called(1);
 
-    verifyNever(cacheManager.getFileStream(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    ));
+    verifyNever(() => cacheManager.getFileStream(
+          url,
+          key: any(named: 'key'),
+          headers: any(named: 'headers'),
+          withProgress: any(named: 'withProgress'),
+        ));
   }, skip: isBrowser);
 
   test('Supplying an CacheManager should call getFileStream', () async {
@@ -78,12 +80,12 @@ void main() {
     ));
     await imageAvailable.future;
 
-    verify(cacheManager.getFileStream(
-      url,
-      withProgress: anyNamed('withProgress'),
-      headers: anyNamed('headers'),
-      key: anyNamed('key'),
-    )).called(1);
+    verify(() => cacheManager.getFileStream(
+          url,
+          key: any(named: 'key'),
+          headers: any(named: 'headers'),
+          withProgress: any(named: 'withProgress'),
+        )).called(1);
   }, skip: isBrowser);
 
   test('Supplying an CacheManager with maxHeight throws assertion', () async {
@@ -101,7 +103,7 @@ void main() {
     result.addListener(
         ImageStreamListener((ImageInfo image, bool synchronousCall) {
       imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace stackTrace) {
+    }, onError: (dynamic error, StackTrace? stackTrace) {
       caughtError.complete(error);
     }));
     final dynamic err = await caughtError.future;
@@ -124,7 +126,7 @@ void main() {
     result.addListener(
         ImageStreamListener((ImageInfo image, bool synchronousCall) {
       imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace stackTrace) {
+    }, onError: (dynamic error, StackTrace? stackTrace) {
       caughtError.complete(error);
     }));
     final dynamic err = await caughtError.future;
